% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{calc_centralize_matrix}
\alias{calc_centralize_matrix}
\title{Function to centralize a matrix X by column}
\usage{
calc_centralize_matrix(X)
}
\arguments{
\item{X}{a matrix. In our package, X is the genotype likelihood matrix, one column for each snp}
}
\value{
A volumn-centeralized matrix for the given matrix X
}
\description{
This function will centeralized each column of the input X by the mean of that column calculated from non-missing data. It is called by \code{\link{sample_bootstrap}} and \code{\link{calc_ScoreV_RVS}}
}

