% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_gener_seqdata.R
\name{calc_EG_general}
\alias{calc_EG_general}
\title{More general function to calculate the conditional expected  genotype probability \eqn{E(P(G_{ij}|D_{ij}))} given the genotype likelihoods \eqn{P(D_{ij}|G_{ij}=g)} and frequencies, see \code{\link{calc_EG}} for simplified version.}
\usage{
calc_EG_general(M, p, rdv)
}
\arguments{
\item{M}{genotype likelihoods with dimension of number of sample by 3 (double), each column is the genotype likelihood \eqn{P(D_{ij}|G=AA, Aa\  or\ aa)} for one locus; uses output from \code{\link{calc_pobs_ndepth}} for simulation data.}

\item{p}{genotype frequencies \eqn{P(G=AA, Aa\ or\ aa)} for each variant; output from function \code{\link{calc_EM}}.}

\item{rdv}{read depth for all samples, should be a vector of integer with length equal to number of samples.}
}
\value{
a vector of containing conditional expected genotype probability. Length should be the same as rdv.
}
\description{
Calculate the conditional expected genotype probability given the observed sequencing data \eqn{D_{ij}} from formula \eqn{E(P(G_{ij}|D_{ij}))=\sum_{g=0}^2 gP(G_{ij}=g|D_{ij})}, where \eqn{ P(G_{ij}=g|D_{ij})=P(D_{ij}|G_{ij}=g)*P(G_{ij}=g)/P(D_{ij})}. It is called from \code{\link{generate_seqdata}}.
}

