% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_gener_seqdata.R
\name{calc_EG_Var}
\alias{calc_EG_Var}
\title{Function to calculate the variance of conditional expected genotype probability \eqn{Var(E(P(G_{ij}|D_{ij})))}.}
\usage{
calc_EG_Var(M, p, rdv)
}
\arguments{
\item{M}{genotype likelihoods \eqn{P(D_{ij}|G=AA, Aa\ or\ aa)} for one locus, matrix of dimension number of samples by 3 (double);}

\item{p}{genotype frequencies \eqn{P(G=AA, Aa\ or\ aa)} (double); output from function \code{\link{calc_EM}}.}

\item{rdv}{read depth (vector of integers) for all samples}
}
\value{
the variance of \eqn{E(P(G_{ij}|D_{ij}))}.
}
\description{
This is the variance for the simplified function \code{\link{calc_EG}}. It used the formula \eqn{var(X)=E(X^2)-E(X)^2}, where \eqn{E(X)} is the same as the result of \code{\link{calc_EG}}.
}

