% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_gener_seqdata.R
\name{calc_EM}
\alias{calc_EM}
\title{Function to use EM algorithm to estimate the genotype frequencies in the sample}
\usage{
calc_EM(M)
}
\arguments{
\item{M}{matrix of genotype likelihoods derived from Phred-scale likelihood in VCF input for one variant.}
}
\value{
a vector with three decimals to stands for probability of 0, 1 or 2 minor alleles.
}
\description{
Given a matrix (M) with dimension number of sample by 3 consisting of genotype likelihoods for each individual, this function uses the EM algorithm to estimate genotype frequencies.
 It returns a vector with three decimals to stands for probability of 0, 1 or 2 minor alleles in the whole sample
}

