% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_gener_seqdata.R
\name{calc_Mr}
\alias{calc_Mr}
\title{Function to calcualte likelihood \eqn{P(D|G)} for simulating data, more general case of \code{\link{calc_pobs_ndepth}}}
\usage{
calc_Mr(error, vect_row_reads, ndepth = 1, rdv)
}
\arguments{
\item{error}{vector of base call error rates; Uses the same error input in the \code{seq_call} function but concatenate all error values over all samples.}

\item{vect_row_reads}{row vector of sequence reads consisting of 'A','G','T' or 'C', same as output from function \code{seq_call}, but concatenated over all samples in same fashion as error.}

\item{ndepth}{read depth (integer), it eaquals to the length of error or vect_row_reads, or sum of vector rdv}

\item{rdv}{integer vector of read depths concatenated for all samples.}
}
\value{
a matrix (M) with 3 columns representing  P(read|AA), P(read|Aa) and P(read|aa), where row dimensions equal length(error)=length(vect_row_reads).
}
\description{
This function calculates a likelihood matrix \eqn{P(D|G)} given a vector of error rates and a vector of sequence reads, where D is the sequencing data and G is the true genotype. It uses function \code{\link{calc_psingle}} to calculate the likelihood for each read given one possible genotype (AA, Aa or aa).
It is a more general case of  \code{\link{calc_pobs_ndepth}} with an additional param rdv, and called from \code{\link{generate_seqdata}}.
}

