% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_gener_seqdata.R
\name{calc_pobs_ndepth}
\alias{calc_pobs_ndepth}
\title{Function to calculate the likelihood matrix \eqn{P(D|G)} given a vector of error rates and a vector of sequence reads.}
\usage{
calc_pobs_ndepth(Error, vect_row_reads, ndepth = 1)
}
\arguments{
\item{Error}{vector of sequencing error rates. Similarly as error in the function \code{\link{seq_call}}, but concatenate all error values over all samples.}

\item{vect_row_reads}{row vector of sequence reads consisting of 'A','G','T' or 'C', similar as output from function \code{\link{seq_call}}, but concatenated over all samples in same fashion as Error.}

\item{ndepth}{integer, read depth, should have value  the same as length of error or length of vect_row_reads.}
}
\value{
a matrix with 3 columns representing P(read|AA), P(read|Aa) and P(read|aa), where row numbers equal to ndepth.
}
\description{
This function calculates a likelihood matrix \eqn{P(D|G)} given a vector of error rates and a vector of sequence reads, where D is the sequencing data and G is the true genotype. It uses function \code{\link{calc_psingle}} to calculate the likelihood for each read given one possible genotype (AA, Aa or aa).
 It is called from function \code{\link{generate_seqdata}}.
}

