% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_gener_seqdata.R
\name{calc_psingle}
\alias{calc_psingle}
\title{Function to  compute the genotype likelihood of a read given the sequencing error rate \eqn{e} and true genotype \eqn{G_1G_2}: \eqn{P(D=g|G=G_1G_2,e), g, G_1, G_2 } have values of A, C, G or T.}
\usage{
calc_psingle(sing_read, A1, A2, error)
}
\arguments{
\item{sing_read}{single read, e.g. A.}

\item{A1}{allele 1 in the true genotype, e.g. if \eqn{G_1G_2=AC} then A1='A'}

\item{A2}{allele 2 in the true genotype, e.g. A2='C'}

\item{error}{sequencing error, decimal with value between 0 and 1.}
}
\value{
likelihood of observing sing_read given genotype (A1A2) and error rate
}
\description{
This function uses the formula \eqn{P(D=g|G=G_1G_2)=1/2*P(D=g|G_1)+1/2*P(D=g|G_2)} and \eqn{p(D=g|G_1=g)=1-e,p(D=other\ call|G_1=g)=e/3} 
to compute the genotype likelihood  \eqn{P(D=g|G=G_1G_2,e)}, more details are in  Appendix B of the paper. It is called from  functions \code{\link{calc_pobs_ndepth}} and \code{\link{calc_Mr}}.
}

