% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{calc_robust_Var}
\alias{calc_robust_Var}
\title{Function to calcualte the robust variance of \eqn{E(Gij|Dij)} for case}
\usage{
calc_robust_Var(P)
}
\arguments{
\item{P}{the genotype frequency, calcualted from EM algorithm}
}
\value{
the Robust variance of E(Gij|Dij)
}
\description{
Use formule \eqn{Var(X)=E(X^2)-E(X)^2} to calculate the variance of genotypes. It is called from \code{\link{RVS_asy}},\code{\link{RVS_btrap}}, and \code{\link{calc_ScoreV_RVS}}.
}

