% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{calc_score_test}
\alias{calc_score_test}
\title{Function to calculate the score test for given expected genotype probability for case and controls seperately (M1, M2)}
\usage{
calc_score_test(M1, M2)
}
\arguments{
\item{M1}{a vector of the expected  genotype probability for case \eqn{E(G_{ij}|D_{ij})} on one snp}

\item{M2}{a vector of the expected genotype probability for control \eqn{E(G_{ij}|D_{ij})} on one snp}
}
\value{
the score test statistic calculated from M1 and M2
}
\description{
This function calcuates the score test \eqn{T=S^2/var(S)}, for a given \eqn{j, S_j=\sum_i (Y_i-bar(Y))E(G_{ij}|D_{ij})=\sum_{case}(1-bar(Y))E(G|D)-\sum_{cont}bar(Y)E(G|D)}. It is called in \code{\link{regScore_perm}}.
}

