% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{calc_score_test1}
\alias{calc_score_test1}
\title{clac_score_test1 calculates the score test for given expected genotype probability for cases and controls seperately (M1, M2)}
\usage{
calc_score_test1(M1, M2)
}
\arguments{
\item{M1}{-  the expected value of genotype for case E(G_ij|D_ij) on one snp (dimension: ncase*1)}

\item{M2}{-  the expected value of genotype for control E(G_ij|D_ij) on one snp (dimension: ncont*1)}
}
\value{
- the score test statistic calculated from M1 and M2 and empirical variance estimation seperated for case and controls.
}
\description{
calc_score_test1 is called in \code{\link{regScore_perm}}
}

