% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{calc_ScoreV}
\alias{calc_ScoreV}
\title{calc_ScoreV gets score vector S for nsnp number of rare variants}
\usage{
calc_ScoreV(X, Y)
}
\arguments{
\item{X}{the expected conditional genotype probability}

\item{Y}{the phenotype}
}
\value{
the score statistic vector S for nsnp rare variant
}
\description{
it is called in \code{regScore_rare}, \code{RVS_rare}
}

