% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{calc_ScoreV_regVar}
\alias{calc_ScoreV_regVar}
\title{calc_ScoreV_regVar: Get Variance of vector S by regular method}
\usage{
calc_ScoreV_regVar(X, Y)
}
\arguments{
\item{X}{-  the expected genotype probability}

\item{Y}{-  the phenotype}
}
\value{
the variance of score statistic vector S by regular method
}
\description{
it is called in \code{regScore_rare}
}

