% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{calc_ScoreV_RVS}
\alias{calc_ScoreV_RVS}
\title{calc_ScoreV_RVS calculates variance of vector S by RVS using robust variance}
\usage{
calc_ScoreV_RVS(X, Y, P, hom, multiplier, snp_loop)
}
\arguments{
\item{X}{the expected conditional genotype probability for case and controls}

\item{Y}{the phenotype}

\item{P}{genotype frequency for J variants P(G=0), P(G=1) and P(G=2), J by 3 matrix}
}
\value{
the variance of score statistic vector S by  RVS using robust variance
}
\description{
it is called in \code{RVS_rare}
}

