% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{check_hom_matrix}
\alias{check_hom_matrix}
\title{Function to remove the homozygosity of a rare SNP in case or control}
\usage{
check_hom_matrix(X, hom = 1, multiplier = 1)
}
\arguments{
\item{X}{the expected condidtional genotype probablity matrix for grouped rare variants}

\item{hom}{with value 1 or 2. 1 means making changes with the 1st nonNA element, 2 means making changes with a random nonNA element}

\item{multiplier}{with value 1 or 2, 1 is dividing by 2 and 2 is mulitpling by 2.}
}
\value{
newX the edited expected condidtional genotype probablity matrix
}
\description{
Given the original case or control genotype matrix, return the edited matrix
}

