% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_parellel.R
\name{clist}
\alias{clist}
\title{clist can combine two lists with the same structure.}
\usage{
clist(list1, list2)
}
\arguments{
\item{list1, }{the first list, each element is a vector}

\item{list2, }{the second list, each element is a vector}
}
\value{
a list with the same structure of list1 or list2, each element is the concatenation of the corresponding element from list1 and list2
}
\description{
clist can combine two lists with the same structure.
}

