% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{combine_twogeno}
\alias{combine_twogeno}
\title{Function to combine two sets of ped pairs for two samples}
\usage{
combine_twogeno(file1, file2, keep1 = NA, keep2 = NA, missing_cut = 0.5)
}
\arguments{
\item{file1}{the file name for the first sample before the extension. Eg. file1=sample1 if the file names are sample1.ped/map}

\item{file2}{the file name for the second sample before the extension. Eg. file1=sample2 if the file names are sample2.ped/map}

\item{keep1}{if only a subset of sample1 is needed in the final analyis, keep1=1, ow, keep1=NA}

\item{keep2}{if only a subset of sample2 is needed in the final analyis, keep2=1, ow, keep2=NA}

\item{missing_cut}{the missing rate to remove the high-missing variants}
}
\value{
a list include data sets 'common' and 'rare' for common  or rare variants and 'snp.Nomiss' for the SNPs kept in common and rare.
}
\description{
Function to combine two sets of ped pairs for two samples
}

