% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{common_coord}
\alias{common_coord}
\title{Function to find the variants presenting in two datasets.}
\usage{
common_coord(cord_set1, cord_set2)
}
\arguments{
\item{cord_set1}{the coordinate of the variants in file1}

\item{cord_set2}{the coordinate of the variants in file2}
}
\value{
a list include two vectors for the location of variants presenting in both files in file1 (cord1) and  file2 (cord2).
}
\description{
This function try to find the variants presenting in two datasets, it will return their location index. Variants presenting in both dataset but with duplicated record will not be returned.
}

