% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{convert}
\alias{convert}
\title{convert two columns of a variant from ped file into additive coding}
\usage{
convert(M, ref)
}
\arguments{
\item{M}{2 columns of a  variant from ped file}

\item{ref}{the first allele for the  variant in map file}
}
\value{
a vector representing additive coding with length the same as sample size for that variant.
}
\description{
convert two columns of a variant from ped file into additive coding
}

