% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{cor.X.function}
\alias{cor.X.function}
\title{Function to edit the correlation matrix of expected genotype probability. It either remove the NA correlation pairs or remove the same pairs (complete LDs).}
\usage{
cor.X.function(X1, hom = 1, multiplier = 1, snp_loop)
}
\arguments{
\item{hom}{with value 1 or 2. 1 means making changes with the 1st nonNA element, 2 means making changes with a random nonNA element}

\item{multiplier}{with value 1 or 2, 1 is dividing by 2 and 2 is mulitpling by 2.}

\item{snp_loop}{will tell the relative location of the probablematic snp.}

\item{X}{the expected condidtional genotype probablity matrix for grouped rare variants}
}
\value{
cor.X the edited correlation from the expected condidtional genotype probablity matrix
}
\description{
Given the original case or control genotype matrix, return the edited correlation matrix. In the check_hom_matrix, we are trying to solve the problem that the whole column is homozygous.
But this NA correlation pairs happens when one column is not homozygous, but when paired with another column, the non-missing values might be homozygous. And it will return the edited correlation matrix
while check_hom_matrix return the edited genotype probability matrix.
}

