% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{filter_by_miss}
\alias{filter_by_miss}
\title{Function to filter out variants by missing rate p}
\usage{
filter_by_miss(M1, M2, missing_cut)
}
\arguments{
\item{M1}{the additive coding of genotype for sample1}

\item{M2}{the additive coding of genotype for sample2}

\item{missing_cut}{the missing rate cut-off, only keep variants with missing rate <p in both samples}
}
\value{
a list including row combined genotype from both sample (M)  and a vector of index whether a variant pass the missingness filter (k).
}
\description{
It removes all variants that missing rate are smaller than p in either sample.
}

