% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{filter_genocall}
\alias{filter_genocall}
\title{Function to filter out variants in the genotype hard call file}
\usage{
filter_genocall(geno, snps, missing_cut = 0.5, maf_cut = 0.05, common = T)
}
\arguments{
\item{geno}{additive coding genotype files, it is the .raw file in plink with 6th columns are phenotype data (output of geno_add_2pheno if no phenotype)}

\item{snps}{a matrix of the variant informations including at least chr, position, alternative and reference alleles}

\item{missing_cut}{a decimal, the missing rate cut off}

\item{maf_cut}{a decimal, the minor allele cut off}

\item{common}{a boolean variable, common = T, return the variants with maf>=maf_cut, ow, return variants with maf<maf_cut.}
}
\value{
a list including- geno: a matrix with the genotype data only (size of nsample*nsnp), snps: matrix with variant information, Y: a vector of phenotype corresponding to the order of the sample in geno, rm_call0: a vector including the position of variants with hom reference calls and rm_miss: a vector including the position of variants with missing_rate>missing_cut in snps after removing rm_call0.
}
\description{
It will filtler out variants by (1) duplicated position, (2) homozygous genotype for all samples, (3) missing rate greater than given missing-cut and (4) given maf.
}

