% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_parellel.R
\name{generate_seq_1ind}
\alias{generate_seq_1ind}
\title{Generate sequence data for one indiviudal with read depth rd 
it calls functions \code{seq_call} to generate sequence data with read depth = rd, each read depth has error error[i].}
\usage{
generate_seq_1ind(maf, rd, error)
}
\arguments{
\item{maf, }{a numeric to specify the maf for one variant}

\item{rd, }{a numeric, read depth of the sequence}

\item{error, }{a vector with length rd, error rate for each squence call}
}
\value{
a list including read depth (rdepth), error rate (error) and a vector of length rdepth showing the sequencing data (seq)
}
\description{
Generate sequence data for one indiviudal with read depth rd 
it calls functions \code{seq_call} to generate sequence data with read depth = rd, each read depth has error error[i].
}

