% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_parellel.R
\name{generate_seqdata_1var}
\alias{generate_seqdata_1var}
\title{Generate sequence data for ncase and ncontrol indiviudal at one loci
it calls functions \code{generate_seq_1ind} to generate sequence data for ncase and ncont at one loci, read depth for two samples (case,cont) follow 
normal distribution of N(mdcase,sdcase) and N(mdcont, sdcont) respectively.}
\usage{
generate_seqdata_1var(ncase, ncont, mdcase, sddcase, mdcont, sddcont, mecase,
  sdecase, mecont, sdecont, mafcase, mafcont)
}
\arguments{
\item{ncase, }{an integer to specify how many case samples}

\item{ncont, }{an integer to specify how many control samples}

\item{mdcase, }{the mean read depth for case}

\item{sddcase, }{the stand variance of the read depth for case}

\item{mdcont, }{the mean read depth for cont}

\item{sddcont, }{the stand deviation of the read depth for cont}

\item{mecase, }{the mean of error rate, probability that sequence call is wrong (double) for case>0}

\item{sdecase, }{standard deviation of error rate (double) for case >0}

\item{mecont, }{the mean of error rate, probability that sequence call is wrong (double) for cont >0}

\item{sdecont, }{standard deviation of error rate (double) for cont >0}

\item{mafcase, }{a numeric to specify the minor allele freq for the variant in case}

\item{mafcont, }{a numeric to  specify the minor allele freq for the variant in cont}
}
\value{
a list including four vectors:  concatenated read depth for each individual (vrdepth, leng(vrdepth)=ncase+ncont), error rate of each sequence for each sample (verror, leng(verror)=sum(vrdepth)), the sequencing data (seq, leng=leng(verror)) and true genotype data (vgeno,leng=leng(vrdepth)),
}
\description{
Generate sequence data for ncase and ncontrol indiviudal at one loci
it calls functions \code{generate_seq_1ind} to generate sequence data for ncase and ncont at one loci, read depth for two samples (case,cont) follow 
normal distribution of N(mdcase,sdcase) and N(mdcont, sdcont) respectively.
}

