% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_seqdata.R
\name{generate_seqdata_OR}
\alias{generate_seqdata_OR}
\title{Function to simulate data  for one variant by given parameters}
\usage{
generate_seqdata_OR(N, preval, ncase, ncont, mafco, OR, mdcase, sdcase, mdcont,
  sdcont, me, sde, nsnp)
}
\arguments{
\item{N}{population size}

\item{preval}{prevalence rate for the disease}

\item{ncase}{the number of case}

\item{ncont}{the number of controls}

\item{mafco}{MAF for controls}

\item{OR}{the odds ratio}

\item{mdcase}{average read depth in cases (double) >0#' @param sddcase, standard deviation of read depth in cases  (double) >0}

\item{mdcont}{average read depth in controls (double) >0}

\item{sde}{standard deviation of error rate (double) >0}

\item{nsnp}{number of variant}

\item{Ncont}{number of controls}

\item{sddcont}{standard deviation of read depth in controls (double) >0#' @param me, average error rate, probability that sequence call is wrong (double) >0}
}
\value{
MM - conditional expected value E(Gij|Dij)

P -estimated genotype frequencies P(G=0), P(G=1), P(G=2)

G - true genotype from which sequence data generated
}
\description{
Function to simulate data  for one variant by given parameters
}

