% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{geno_add_2pheno}
\alias{geno_add_2pheno}
\title{Function to add the phenotype in the 6th column of the genotype file recognizable by plink, eg, file with extension 'ped' or 'raw'.}
\usage{
geno_add_2pheno(genofile, caseIDfile)
}
\arguments{
\item{genofile}{additive coding genotype files, it is the .raw file in plink}

\item{caseIDfile}{the file including the list of case IDs}
}
\value{
a list including- geno: the genotype data with phenotype on 6th column, snps: variant information in bim format (in plink) and ncase: number of case
}
\description{
Function to add the phenotype in the 6th column of the genotype file recognizable by plink, eg, file with extension 'ped' or 'raw'.
}

