% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{geno_process}
\alias{geno_process}
\title{Function to deal with genotype hard call files in additive coding}
\usage{
geno_process(genofile, caseIDfile, missing_cut, maf_cut, common)
}
\arguments{
\item{genofile}{additive coding genotype files, it is the .raw file in plink}

\item{caseIDfile}{the file including the list of case IDs}

\item{missing_cut}{a decimal, the missing rate cut off}

\item{maf_cut}{a decimal, the minor allele cut off}

\item{common}{a boolean variable, common = T, return the variants with maf>=maf_cut, ow, return variants with maf<maf_cut.}
}
\value{
a list including- geno: a matrix with the genotype data only (size of nsample*nsnp), snps: matrix with variant information, Y: a vector of phenotype corresponding to the order of the sample in geno, rm_call0: a vector including the position of variants with hom reference calls and rm_miss: a vector including the position of variants with missing_rate>missing_cut in snps after removing rm_call0.
}
\description{
It will read the additive coding genotype file, filter out variants and then return the phenotype and genotype file necessary for association test.
It calls functions \code{\link{geno_add_2pheno}} and \code{\link{filter_genocall}}
}

