% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{get_allele_frq}
\alias{get_allele_frq}
\title{Function to calculate the allele frequency from additive coding genotype}
\usage{
get_allele_frq(M)
}
\arguments{
\item{M}{the additive coding genotype, a matrix with column for each variant}
}
\value{
a vector of allele frequency calculated based on the non-missing data
}
\description{
Function to calculate the allele frequency from additive coding genotype, the number of allele is not necessarily for minor allele. For people know plink, there is another option to calcualte the MAF.
}

