% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_vcf.R
\name{get_exp_geno}
\alias{get_exp_geno}
\title{Function to generates the expected probabilities of the genotypes E(P(G_ij|D_ij)).}
\usage{
get_exp_geno(A0M, A1M, A2M, B0M, B1M, B2M, chr, loc)
}
\arguments{
\item{A0M}{Genotype likelihood for major homozygous of the cases, output of filter_VCF_SNPs.}

\item{A1M}{Genotype likelihood for minor heterozygous of the cases, output of filter_VCF_SNPs.}

\item{A2M}{Genotype likelihood for minor homozygous of the cases, output of filter_VCF_SNPs.}

\item{B0M}{Genotype likelihood for major homozygous of the controls, output of filter_VCF_SNPs.}

\item{B1M}{Genotype likelihood for minor heterozygous of the controls, output of filter_VCF_SNPs.}

\item{B2M}{Genotype likelihood for minor homozygous of the controls, output of filter_VCF_SNPs.}

\item{chr}{chromosome of the variant, output from the  function filter_VCF_SNPs}

\item{loc}{location of the variant, output from the function filter_VCF_SNPs.}
}
\value{
a list includes a matrix (exp_cond_prob) contains the expected genotype likelihood for each person (row) at each locus (column); a matrix (pop_freq) for genotype frequency in all the sample at each locus (size: number of variant *3); a data frame (SNPs) contains chr and loc of the returned variants and number of variants (nrm_hom) removed because of homozygous calls in the whole sample.
}
\description{
Using the genotype likelihood for case and controls from VCF file to generates the population frequency by calling function \code{\code{calc_EM}} and then use it to calculate the expected genotype probabilities E(P(G_ij|D_ij)) by calling function \code{\link{calc_EG}}. Variants with homozygous call in the whole sample (the std of their E(P(G_ij|D_ij)) <10^4) will be removed and the expected genotype probability will be reorganized so that the top rows are for case and the rest for controls.
}

