% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_vcf.R
\name{get_exp_MAF}
\alias{get_exp_MAF}
\title{Function to calculate the minor allele frequency (MAF) from conditional expected genotype probability.}
\usage{
get_exp_MAF(P, MM, chr, loc, ncase, maf_cut = 0.05, common = common)
}
\arguments{
\item{P}{genotype frequency for the whole sample which is the output of  function get_exp_geno.}

\item{MM}{expected genotype probability,  output of  function get_exp_geno.}

\item{chr}{the chrom info of each variant}

\item{loc}{the location info of each variant}

\item{ncase}{integer, the number of cases}

\item{maf_cut}{the minor allele frequency cut-off  for common or rare variants}

\item{common}{Boolean variable to indicate common or rare variants}
}
\value{
matrix of genotype frequency (P), expected genotype probability (Geno) and variant information (SNPs), number of case (ncase), number of control (ncont) and number of variants (nsnp) and a vector for phenotyp (Y).
}
\description{
It calculates the MAF from conditional expected genotype probability and returns the matrices of expected genotype probability and genotype frequncy for variants indicated by maf_cut and group which are to be inputed into a test function of the rvs package.
}

