% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_vcf.R
\name{get_IDs}
\alias{get_IDs}
\title{Function to seprates the case and control IDs.}
\usage{
get_IDs(vcf_file, caseID_file, nhead = 128, ncol_ID = 9)
}
\arguments{
\item{vcf_file}{The saving address and filename  of the vcf file in the format of 'address/filename'.}

\item{caseID_file}{The  file including case IDs in the format of 'address/filename'.}

\item{nhead}{The number of the lines in the vcf file which started with #.}

\item{ncol_ID}{The number of columns before the sample IDs start in the last line of the headers in the vcf file, default=9.}
}
\value{
a list which inclues two vectors that correspond to the position index of the cases or controls in the last head row of the VCF with IDs, but columns are shifted by the ncol_ID columns.
}
\description{
This function read the header of the vcf file (rows starting with # in vcf) and split the case and controls according to the caseID_file. It returns two vectors each of which is the position index of case or control in the list of samples in vcf file (the last row starting with #).
}

