% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_vcf.R
\name{get_lsingle_vcf}
\alias{get_lsingle_vcf}
\title{Function to parse Phred-scaled likelihoods}
\usage{
get_lsingle_vcf(Mij, n_pl)
}
\arguments{
\item{Mij}{a single data unit of format'GT:AD:DP:GQ:PL'.}

\item{n_pl}{an integer to indicate the PL field.}
}
\value{
a vector of three values: l0, l1, l2
}
\description{
This function parses the data column (GT:AD:DP:GQ:PL) to get the PL values -10log10(x),
and converts them to a vector of likelihoods: L0,L1,L2. It's called within \code{\link{get_likelihood_vcf}}.
newer version of workflow of GATK suggested calling the BAM for each individual first then combined them, 
so the format column can have different length for each variant like the example Lizhen provided.
}

