% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{minor_allele}
\alias{minor_allele}
\title{Function to find the minor allele in the whole sample}
\usage{
minor_allele(M1, M2)
}
\arguments{
\item{M1}{the 2 column genotype for a variant from sample1's ped file}

\item{M2}{the 2 column genotype for a variant from sample2's ped file}
}
\value{
the minor allele from the two inputed samples (character).
}
\description{
Function to find the minor allele in the whole sample
}

