% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_gener_seqdata.R
\name{odds.2.probs}
\alias{odds.2.probs}
\title{Function to calculate the probability from given odds}
\usage{
odds.2.probs(odds)
}
\arguments{
\item{odds}{odds}
}
\value{
probs probability corresponding to the given odds
}
\description{
Function to calculate the probability from given odds
}

