% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{ped_2_additive}
\alias{ped_2_additive}
\title{Function to combine two ped files for case and controls and transform them into additive coding}
\usage{
ped_2_additive(M1, M2, ncols)
}
\arguments{
\item{M1}{the genotype data for the first sample (normally case)}

\item{M2}{the genotype data for the secondsample (normally controls)}

\item{ncols}{the number of column before the genotype data start}
}
\value{
A list including - Ad_all: the additive coding for all snps left, ncase: the first ncase rows are for cases;
 minor: the minor allele for each variants; rmed, the index of the removed snps from the input
}
\description{
This function remove variants that are missing in one sample, monomorphic in whole sample, monomorphic in each sample with diff alleles, non biallelic
 Alternatively, for those know plink, all these can be done by plink to merge the two files and generate the additive coding with option --recodeA
}

