% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_parellel.R
\name{rbindlist}
\alias{rbindlist}
\title{rbindlist can rbind combine two lists with the same structure.}
\usage{
rbindlist(list1, list2)
}
\arguments{
\item{list1, }{the first list,each element is a vector}

\item{list2, }{the second list,each element is a vector}
}
\value{
a list,  each element is the rbinding of the corresponding element from list1 and list2
}
\description{
rbindlist can rbind combine two lists with the same structure.
}

