% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{regScore_perm}
\alias{regScore_perm}
\title{Use regular Score/Trend test for association for given phenotype (Y) and expected values of genotypes for case and controls (X)
it is evaluation by permutation distribution}
\usage{
regScore_perm(Y, X, nperm)
}
\arguments{
\item{Y}{- a vector with the phenotype for n individuals (=ncase+ncont), first ncase (y=1) then ncont (y=0)  (integar)}

\item{X}{- a vector of genotype for a snp, first ncase and then ncont}

\item{nperm}{- number of permutations}
}
\value{
p-values the variant (double)
}
\description{
Use regular Score/Trend test for association for given phenotype (Y) and expected values of genotypes for case and controls (X)
it is evaluation by permutation distribution
}

