% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{reset_minor_additive}
\alias{reset_minor_additive}
\title{Function to reset the additive coding genotype into nubmer of minor alleles.}
\usage{
reset_minor_additive(M)
}
\arguments{
\item{M}{the additive coding genotype matrix for the whole sample}
}
\value{
a matrix for additive genotype in terms of number of minor allels.
}
\description{
This is not needed if your genotype coding is based on the number of minor allele. It calls \code{\link{get_allele_frq}} for the allele frequency then reset the genotype matrix if frequency >0.5.
}

