% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{RVS_btrap}
\alias{RVS_btrap}
\title{use RVS to test associaton by bootrtrap, given phenotype (Y), expected values of genotypes for case and controls (X)
estimated genotype frequency (P) and number of times of bootstrap (nboot)}
\usage{
RVS_btrap(Y, X, P, nboot, RVS = "TRUE")
}
\arguments{
\item{Y}{- a vector with the phenotype for n individuals (=ncase+ncont), first ncase (y=1) then ncont (y=0)  (integar)}

\item{X}{- a vector of genotype for a snp, first ncase and then ncont}

\item{P}{(vector with length 3  (double)) - estimated genotype frequency for a variant P(G=0), P(G=1) and P(G=2)}

\item{nboot}{- number of bootstrap}

\item{RVS, }{logic variable to indicate the estimation of variance of the score statistic.}
}
\value{
p-values for the snp
}
\description{
use RVS to test associaton by bootrtrap, given phenotype (Y), expected values of genotypes for case and controls (X)
estimated genotype frequency (P) and number of times of bootstrap (nboot)
}

