% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{RVS_rare2}
\alias{RVS_rare2}
\title{It calls functions \code{calc_ScoreV},\code{calc_ScoreV_RVS},\code{calc_ScoreV_likely},\code{test_CAST}, \code{test_Calpha},\code{sample_bootstrap}}
\usage{
RVS_rare2(Y, X, P, njoint = 5, nboot, RVS = "TRUE", hom = 1,
  multiplier = 1, snp_loop = 1)
}
\arguments{
\item{Y}{- phenotype value, 1-cases and 0-controls}

\item{X}{- matrix of conditional expected values, it has dimensions n by J, J - number of variants, n - sample size}

\item{P}{- genotype frequency for J variants P(G=0), P(G=1) and P(G=2), J by 3 matrix}

\item{njoint}{- number of SNPs grouping together for one test, default is 5}

\item{nboot}{- number of bootstrap}

\item{RVS, }{logic variable to indicate the estimation of variance of the score statistic.}
}
\value{
two p-values for CAST and C-alpha
}
\description{
It calls functions \code{calc_ScoreV},\code{calc_ScoreV_RVS},\code{calc_ScoreV_likely},\code{test_CAST}, \code{test_Calpha},\code{sample_bootstrap}
}

