% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_gener_seqdata.R
\name{seq_call}
\alias{seq_call}
\title{Function to produce a sequencing read at one locus with converage=ndepth from given a genotype and the error scores (one error for one read).}
\usage{
seq_call(genotype, error, ndepth)
}
\arguments{
\item{genotype}{vector of two alleles for a given genotype, ie. genotype CT is given as ('C', 'T')}

\item{error}{vector of sequencing errors, length of this vector is the same as read depth (ndepth).}

\item{ndepth}{the read depth (integer)}
}
\value{
a row vector of sequence reads consisting of 'A','G','T' or 'C' for a single variant with length the same as read depth (ndepth).
}
\description{
It uses formula \eqn{p(D=g|G_1=g)=1-e,p(D=other\ call|G_1=g)=e/3}  to generate the read for a given genotype \eqn{G_1} and error rate \eqn{e}.
It is called from the \code{\link{generate_seqdata}} functions.
}

