% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_parellel.R
\name{seq_eg_freq}
\alias{seq_eg_freq}
\title{use the simulated data to calculate the expected genotype probability and population frequency}
\usage{
seq_eg_freq(vgeno, seq, verror, vrdepth)
}
\arguments{
\item{vgeno, }{the concatenated true genotype for each sample}

\item{seq, }{the concatenated sequence data}

\item{verror, }{the concatenated error rate for each sequence data}

\item{vrdepth, }{the concatenated error rate for each sample}
}
\value{
a list including expected genotype probability (exp_prob, leng=nsample), population fequency (pop_frq,leng=3),true genotype(geno),and read depth(vrdepth), two numerics for number of cases (ncase) and controls (ncont)
}
\description{
use the output of  \code{generate_seqdata_1var} as input to calculate the expected genotype probability and population frequency
It calls subfunction \code{calc_Mr} and \code{calc_EM} for population frequency of P(G=AA,AB,BB),
It calls subfunction \code{calc_pobs_ndepth}, \code{calc_EG_general} and the populationfrequency for expected genotype probability
}

