% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_parellel.R
\name{seqdata_nvar}
\alias{seqdata_nvar}
\title{Generate sequence data for nsnp variants}
\usage{
seqdata_nvar(nsnp, ncase, ncont, mdcase, sddcase, mdcont, sddcont, mecase,
  sdecase, mecont, sdecont, mmafcase, mmafcont)
}
\arguments{
\item{nsnp, }{number of variants (integer) >0}

\item{ncase, }{number of cases (integer) >0}

\item{ncont, }{number of controls (integer) >0}

\item{mdcase, }{average read depth in cases (double) >0}

\item{sddcase, }{standard deviation of read depth in cases  (double) >0}

\item{mdcont, }{average read depth in controls (double) >0}

\item{sddcont, }{standard deviation of read depth in controls (double) >0}

\item{mecase, }{average error rate, probability that sequence call is wrong (double) for case>0}

\item{sdecase, }{standard deviation of error rate (double) for case >0}

\item{mecont, }{average error rate, probability that sequence call is wrong (double) for cont >0}

\item{sdecont, }{standard deviation of error rate (double) for cont >0}

\item{mmafcase, }{a vector of MAF for nsnp variants in cases, HWE is assumed >0}

\item{mmafcont, }{a vector of MAF for nsnp variants in controls}
}
\value{
a list including expected genotype probability (exp_prob, leng=nsample), population fequency (pop_frq,leng=3),true genotype(geno),and read depth(vrdepth), and ncase and ncont
}
\description{
It calls functions \code{generate_seqdata_1var},\code{seq_eg_freq},\code{rbindlist} to generate sequence data,
return a list including expected genotype probability (exp_prob, leng=nsample), population fequency (pop_frq,leng=3),true genotype(geno),concatenated error rate(verror) and read depth(vrdepth)
}

