% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_gener_seqdata.R
\name{sim.corr.binary.data}
\alias{sim.corr.binary.data}
\title{Function to simulate the correlated binary data with given OR.}
\usage{
sim.corr.binary.data(N, preval, mafco, OR, ncase, ncont)
}
\arguments{
\item{N}{Population size, greater than ncase+ncont}

\item{preval}{prevalence rate of the disease}

\item{mafco}{MAF for controls}

\item{OR}{the odds ratio}

\item{ncase}{Number of cases in the taken sample, taken from case of the N population}

\item{ncont}{Number of controls in the sample, take from the controls of N population}
}
\value{
a dataframe include genotype data (x) and phenotype data (y) in each column
}
\description{
Function to simulate the correlated binary data with given OR.
}

