% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{subset_ind}
\alias{subset_ind}
\title{Function to take a subset the samples from a larger genotype data}
\usage{
subset_ind(path1, path2)
}
\arguments{
\item{path1}{path and file name of the larger genotype files where subset is taken from, ped file is prefered}

\item{path2}{the file including the list of IDs for individuals to be kept}
}
\value{
a list including the genotype data matrix (geno1) and a interger for the number of samples (n1).
}
\description{
This function will take a subset of samples' genotype data from a larger dataset, only return the genotype for the wanted individuals.
Similar as the option --keep in plink.
}

