% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raw_process.R
\name{subset_snp}
\alias{subset_snp}
\title{Function to extract the genotype columns for chosen variants from a ped format file.}
\usage{
subset_snp(M, set)
}
\arguments{
\item{M}{the genotype file in ped format,Note, each snp have two columns in the ped file and the first 6 column need to be included.}

\item{set}{the return of common_coord, the vector of integers representing the location of the common snps in the two sets.}
}
\value{
the genotype file for the snps common in two sets only, each snp has two columns.
}
\description{
Given the full set of genotype file in plink format and variants index would like to keep, this function will return the genotype columns for the wanted variants (the first 6 columns are removed). Similar to the option --extract in plink software, but the first 6 columns are removed in this function.
}

