% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{test_Calpha}
\alias{test_Calpha}
\title{Function to obtain p-value by Calpha method (quadratic test) for J rare variant}
\usage{
test_Calpha(S, Sigma)
}
\arguments{
\item{S}{vector of length J which includes the Scores for each variant}

\item{Sigma}{the variance matrix of vector S.}
}
\value{
p-values from Calpha test for J variants (double)
}
\description{
it is called in \code{\link{regScore_rare}} and \code{\link{RVS_rare}}
}

