% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{test_CAST}
\alias{test_CAST}
\title{Function to obtain p-value by CAST method (linear test) for J rare variant}
\usage{
test_CAST(S, Sigma, w)
}
\arguments{
\item{S}{vector of length J which includes the Scores for each variant}

\item{Sigma}{the variance matrix of vector S.}

\item{w}{the weight for CAST test}
}
\value{
p-values from CAST test for J variants (double)
}
\description{
it is called in \code{regScore_rare} and \code{RVS_rare}
}

