% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_assoc_test.R
\name{test_Hotelling}
\alias{test_Hotelling}
\title{Function to obtain pvalue by Hotelling test for J number of rare variants}
\usage{
test_Hotelling(S, Sigma)
}
\arguments{
\item{S}{vector of length J which includes the Scores for each variant}

\item{Sigma}{the variance matrix of vector S.}
}
\value{
p-values from Hotelling test for J variants (double)
}
\description{
Function to obtain pvalue by Hotelling test for J number of rare variants
}

