% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_vcf.R
\name{vcf_process}
\alias{vcf_process}
\title{the main function to call the VCF file and return the  expected probabilities of the genotypes E(G_ij|D_ij) from VCF file}
\usage{

  vcf_process(file = "example/1g113low_1g56exomehigh_filtered.hg19.chr11_1000snps.vcf",
  caseID_file = "example/caseID.txt", nhead = 128, ncol_ID = 9,
  missing_th = 0.2, nvar_tot = 1000, nread = 300, maf_cut = 0.05,
  common = TRUE)
}
\arguments{
\item{file}{the VCF file}

\item{caseID_file}{the caseID files}

\item{nhead}{the row number of your VCF which contains all sample ID}

\item{ncol_ID}{the number of columns before the sample ID}

\item{missing_th}{the missing rate cut off for a SNP to be filtered, if missing>missing_th in case or controls, the SNP will be removed}

\item{nvar_tot}{should be the number of SNPs in your VCF file, but if VCF is too large, you can choose only read in the first nsnp SNPs}

\item{nread}{the number of SNPs read each time by R (nread<=nsnp)}

\item{maf_cut}{the MAF to be used to filter out SNP,}

\item{common}{Boolean variable to indicate common (TRUE) or rare (FALSE) variants}
}
\value{
a list  for case or control (depends on group) including expected conditional probabilities, SNP info, Y, ncase/ncont etc
}
\description{
the main function to call the VCF file and return the  expected probabilities of the genotypes E(G_ij|D_ij) from VCF file
}

